% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{bedload}
\alias{bedload}
\title{Calculate sediment transport rate}
\usage{
bedload(R, D50, D84, S, Tc50 = 0.036)
}
\arguments{
\item{R}{hydraulic radius for channel cross section (m)}

\item{D50}{median surface grain size (mm)}

\item{D84}{84th percentile of bed surface in main channel (mm)}

\item{S}{reach average channel gradient (m/m)}

\item{TC50}{(optional) critical dimensionless stress for D50 (default = 0.036)}
}
\description{
\code{bedload} estimates the bedload transport rate per unit width based on 
the Eaton and Church (2011) function, using the bed surface D50 and the mean
 boundary shear stress
}
