% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{crit_stress}
\alias{crit_stress}
\title{Estimate the critical shear stress}
\usage{
crit_stress(Di, D50, TC50 = 0.045)
}
\arguments{
\item{Di}{vector of grain sizes to analyse (mm)}

\item{D50}{median grain size for the bed surface (mm)}

\item{TC50}{(optional) critical dimensionlessstress for D50 (default is 0.045)}
}
\description{
\code{crit_stress} estimates the shear stress required to entrain a grain of
size Di from a bed surface with a median size of D50. The equations are based
on the equations presented in Wilcock and Crowe (2001), assuming a critical
dimensionless shear stress for the D50 of 0.045 (as a default)
}
