% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{est_effective}
\alias{est_effective}
\title{Plot sediment transport data from stochasim}
\usage{
est_effective(results, Map = F)
}
\arguments{
\item{results}{a data frame containing the output from \code{run_stochasim}}

\item{Map}{(optional) LOGICAL, when TRUE, the transp. frequency dist. is graphed}
}
\description{
\code{est_effective} takes results produced by running STOCHASIM and
estimates the effective discharge (i.e. the one that does the most geom
-orphic work. The function can also produce a plot of the proportion of 
the total work associated with a given flow.
}
