% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{est_formative}
\alias{est_formative}
\title{Estimate the formative discharge for a channel based on bank erosion}
\usage{
est_formative(results, Map = F)
}
\arguments{
\item{results}{a data frame containing the output from \code{run_stochasim}}

\item{Map}{(optional) LOGICAL, when TRUE, the bank erosion frequency dist. is graphed}
}
\description{
\code{est_formative} takes results produced by running STOCHASIM and
estimates the formative discharge, defined to be the discharge that does the
most bank erosion over time. Smaller flows can erode the banks, but these 
ones are responsible for the majority of the bank erosion
}
