% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{est_mu}
\alias{est_mu}
\title{Approximate relative bank strength}
\usage{
est_mu(H, d)
}
\arguments{
\item{H}{effective rooting depth for riparian vegetation (m)}

\item{d}{mean hydraulic depth at channel-forming flow (m)}
}
\description{
\code{est_mu} uses an empirical function to estimate the relative bank
strength produced by vegetation with a known effective rooting depth (H)
(after Eaton, 2006) and a known average channel depth (d). The function only
applies when H < 0.9d.
}
