% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydist.R
\name{hydist}
\alias{hydist}
\title{Create a new hydrograph distribution}
\usage{
hydist(realise)
}
\arguments{
\item{realise}{Function that accepts a single non-negative integer and
returns that many \code{gbem::hydrograph()} realizations.
hydrograph.}
}
\value{
A hydrograph distribution.
}
\description{
Create a new hydrograph distribution
}
