% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydist_rain.R, R/hydist_snow.R
\name{hydist_rain}
\alias{hydist_rain}
\alias{hydist_snow}
\title{Make a Rainfall or Snowmelt-Related Event Hydrograph Distribution}
\usage{
hydist_rain(dst_peak, baseflow, duration = 3 * 24)

hydist_snow(dst_peak, baseflow, duration = 9 * 24)
}
\arguments{
\item{dst_peak}{Distionary distribution of the peak. Left-truncated so as
to always exceed the baseflow.}

\item{baseflow}{Baseflow discharge; single numeric.}

\item{duration}{Length of the event. Defaults to 3 days for
rainfall-related events, and 9 days for snowmelt-related events (in hours).}
}
\value{
A hydrograph distribution for rainfall-related and snowmelt-related
events, with a constant baseflow and a stochastic peak discharge.
}
\description{
Make a Rainfall or Snowmelt-Related Event Hydrograph Distribution
}
\examples{
d <- distionary::dst_gev(100, 3, 0.1)
set.seed(4)
hs <- hydist_snow(d, 50)
plot(hs, n = 1000, col = "orange3")
hr <- hydist_rain(d, 50)
plot(hr, n = 1000, col = "blue3", add = TRUE)
}
