% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_floods.R
\name{make_floods}
\alias{make_floods}
\title{Generate a sequence of floods to run the simulations based on the
specified GEV distributions}
\usage{
make_floods(gev, n_floods)
}
\arguments{
\item{gev}{Specify a GEV}

\item{n_floods}{description}
}
\description{
Generate a sequence of floods to run the simulations based on the
specified GEV distributions
}
