% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydist.R
\name{new_hydist}
\alias{new_hydist}
\title{Constructor Function for "hydst" Objects}
\usage{
new_hydist(l, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a distribution object.}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\value{
\code{"hydist"} object.
}
\description{
Constructor Function for "hydst" Objects
}
