% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim.R
\name{new_stochasim}
\alias{new_stochasim}
\title{Constructor Function for Stochasim Objects}
\usage{
new_stochasim(l, ..., class = character())
}
\arguments{
\item{l}{List to convert to stochasim object.}

\item{...}{Attributes to add to the object.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for Stochasim Objects
}
