% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hydist.R
\name{plot.hydist}
\alias{plot.hydist}
\title{Plot random realization of a hydrograph distribution}
\usage{
\method{plot}{hydist}(x, ..., ylab = "Discharge", xlab = "Time")
}
\arguments{
\item{x}{Hydrograph distribution}

\item{...}{Other arguments to pass to the curve() function.}

\item{xlab, ylab}{Y-axis and x-axis labels for the plot.}
}
\value{
Base R plot of randomly drawn hydrograph.
}
\description{
Plot random realization of a hydrograph distribution
}
