% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stochasim}
\alias{plot.stochasim}
\title{Plotting a stochasim object}
\usage{
\method{plot}{stochasim}(x, ..., what = c("widths", "flows"), from = 0)
}
\arguments{
\item{x}{Stochasim object, as a result of running \code{stochasim()} or
\code{stochasim2()}.}

\item{...}{Other arguments to pass to the default \code{plot()} function.}

\item{what}{What to plot from the stochasim object. One of \code{"widths"},
for the evolution of channel width, or \code{"flows"}, for peak flows over time.}

\item{from}{Iteration number to begin plotting from.}
}
\description{
Plotting a stochasim object
}
