% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{plot_erosion}
\alias{plot_erosion}
\title{Plot erosion data from stochasim}
\usage{
plot_erosion(results, Est = F)
}
\arguments{
\item{results}{a data frame containing the output from \code{run_stochasim}}

\item{Est}{(optional) LOGICAL, TRUE returns RI of bank erosion}
}
\description{
\code{plot_erosion} takes the results produced by running the STOCHASIM
model and produces a plot of the bank erosion magnitude and frequency
}
