% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{plot_floods}
\alias{plot_floods}
\title{Plot flood frequency distribution}
\usage{
plot_floods(results)
}
\arguments{
\item{results}{a data frame containing the output from \code{run_stochasim}}
}
\description{
\code{plot_floods} takes results produced by running STOCHASIM and
produces a typical flood frequency plot. The data are color coded by 
whether or not the channel was unstable during the flood event. the limits
of the stable and unstable flood populations are indicated with a dashed
line.
}
