% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series of discharge and width}
\usage{
plot_timeseries(results)
}
\arguments{
\item{results}{a data frame containing the output from \code{run_stochasim}}
}
\description{
\code{plot_timeseries} takes results produced by running STOCHASIM and
produces plots of the discharge series and of the channel width series
simulated by the model. The interquartile range for each distribution is
shown on the plots using dashed red lines.
}
