% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydist.R
\name{realise.hydist}
\alias{realise.hydist}
\title{Generate Random Hydrographs}
\usage{
\method{realise}{hydist}(distribution, n = 1)
}
\arguments{
\item{distribution}{Hydrograph distribution.}

\item{n}{Number of hydrographs to generate; single positive integer.}
}
\description{
Generate Random Hydrographs
}
