% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revegetate.R
\name{revegetate}
\alias{revegetate}
\title{Revegetate a cross section}
\usage{
revegetate(cross_section, width0, rate = 0.1)
}
\arguments{
\item{cross_section}{A \code{"cross_section"} object representing a stream's
cross section.}

\item{width0}{"Stable" cross section width; positive numeric.}

\item{rate}{Revegetation percent; single number between 0 and 1.}
}
\value{
The original cross section with a reduced width. Width reduction
is calculated as the proportion \code{rate} of the difference between channel
width and \code{width0}.
}
\description{
Reduce the width of a channel cross section by a percentage
compared to a baseline channel width.
}
