% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{run_stochasim}
\alias{run_stochasim}
\title{Run the STOCHASIM model}
\usage{
run_stochasim(gsd, Q, slope, H = 0, W = NA, D = NA, Qmaf = NA, veg.rate = 0.1)
}
\arguments{
\item{gsd}{a data frame containing information on the bed surface GSD (see \code{sim_gsd})}

\item{Q}{a vector of floods to be used sequentially in the simulation (m3/s)}

\item{slope}{reach average channel slope (m/m)}

\item{H}{(optional) estimated effective rooting depth, affecting bank stability (m)}
}
\description{
\code{run_stochasim} executes the stochastic channel simulator code The
model predicts channel widening when a threshold shear stress is exceeded
and otherwise predicts channel narrowing due to vegetation colonization.
The length of the simulation is determined by the length of the input vector
containing the random sequence of annual peak flows (which can be created
using \code{sim_flood} or generated by the user).
}
