% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{sim_gsd}
\alias{sim_gsd}
\title{Simulate a log-normal grain size distribution}
\usage{
sim_gsd(D50, sp = 1, Map = FALSE)
}
\arguments{
\item{D50}{median bed surface grain size (mm)}

\item{sp}{(optional) spread of the distribution (default = 1)}

\item{Map}{(optional) LOGICAL, when TRUE, the distribution is graphed}
}
\description{
\code{sim_gsd} generates a log-normal grain size distribution for a
user-specified D50. The user can also vary the spread of the distribution.
The function returns a data frame containing the upper range of the size
classes, the mid point of each size class, the proportion in each size class,
and the cumulative proportion finer than the upper size class boundaries.
}
