% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{solve_Q}
\alias{solve_Q}
\title{Solve for water depth}
\usage{
solve_Q(Q, W, S, D84)
}
\arguments{
\item{Q}{discharge contained by the channel (m3/s)}

\item{W}{water surface width for the channel (m)}

\item{S}{reach average channel gradient (m/m)}

\item{D84}{84th percentile of the bed surface grain size distribution (mm)}
}
\description{
\code{solve_Q} uses a midpoint solution approach to find water depth that
produces the user-specified stream discharge in a rectangular channel with
a known channel width.
}
