% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_flows.R, R/ss_widths.R
\name{ss_flows}
\alias{ss_flows}
\alias{ss_widths}
\title{Get peak flows and channel widths from stochasim}
\usage{
ss_flows(stochasim)

ss_widths(stochasim)
}
\arguments{
\item{stochasim}{Stochasim object.}
}
\description{
Get peak flows and channel widths from stochasim
}
