% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ss_plot_flows}
\alias{ss_plot_flows}
\title{Plot flows from stochasim object}
\usage{
ss_plot_flows(stochasim, from, ...)
}
\arguments{
\item{stochasim}{Stochasim object, as a result of running \code{stochasim()} or
\code{stochasim2()}.}

\item{from}{Iteration number to begin plotting from.}

\item{...}{Other arguments to pass to the default \code{plot()} function.}
}
\description{
Plot flows from stochasim object
}
