% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{stochasim_dual}
\alias{stochasim_dual}
\title{Run stochasim using hydrographs and two flood generating mechanisms}
\usage{
stochasim_dual(
  floods_snow,
  floods_rain,
  Q_base,
  n,
  d84,
  d50,
  width_0,
  slope,
  H
)
}
\arguments{
\item{floods_snow}{The output from make_floods}

\item{floods_rain}{The output from make_floods}

\item{Q_base}{Mean annual flow}

\item{n}{Manning's n value for the main channel}

\item{d84}{84th percentile of the surface grain size distribution (mm)}

\item{d50}{50th percentile of the grain size distribution (mm)}

\item{width_0}{Water surface width at the beginning of time interval T}

\item{slope}{Energy gradient of the stream channel}

\item{H}{Effective rooting depth for vegetation}
}
\description{
Run stochasim using hydrographs and two flood generating mechanisms
}
