% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain.R, R/snow.R
\name{stochasim_rain}
\alias{stochasim_rain}
\alias{stochasim_snow}
\title{Function to run stochasim using snow hydrographs}
\usage{
stochasim_rain(floods_rain, Q_base, n, d84, d50, width_0, slope, H)

stochasim_snow(floods_snow, Q_base, cross_section)
}
\arguments{
\item{floods_rain}{The output from make_floods}

\item{Q_base}{Mean annual flow}

\item{n}{Manning's n value for the main channel}

\item{d84}{84th percentile of the surface grain size distribution (mm)}

\item{d50}{50th percentile of the grain size distribution (mm)}

\item{width_0}{Water surface width at the beginning of time interval T}

\item{slope}{Energy gradient of the stream channel}

\item{H}{Effective rooting depth for vegetation}

\item{floods_snow}{The output from make_floods}
}
\description{
Function to run stochasim using snow hydrographs
}
