% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{stress}
\alias{stress}
\title{Calculate the mean boundary shear stress}
\usage{
stress(R, S, rho = 1000)
}
\arguments{
\item{R}{hydraulic radius for channel cross section (m)}

\item{S}{reach average channel gradient (m/m)}

\item{rho}{(optional) fluid density (kg/m3), default density is water}
}
\description{
\code{stress} calculates the mean boundary shear stress (Pa), using the
hydraulic radius, reach average gradient and fluid density
}
