% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim_functions.R
\name{velocity}
\alias{velocity}
\title{Estimate flow velocity}
\usage{
velocity(R, S, D84)
}
\arguments{
\item{R}{hydraulic radius for channel cross section (m)}

\item{S}{reach average channel gradient (m/m)}

\item{D84}{84th percentile of bed surface in main channel (mm)}
}
\description{
\code{velocity} returns the mean flow velocity (m/s) for a given hydraulic
radius, bed surface D84, and reach average gradient using Ferguson (2007)
}
