% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim.R
\name{stochasim}
\alias{stochasim}
\title{Stochasim Algorithm}
\usage{
stochasim(hydist, cross_section, rv_rate = 0.1, nsim = 1000, niter = 300)
}
\arguments{
\item{hydist}{Hydrograph Distribution}

\item{cross_section}{A \code{"cross_section"} object representing a stream's
cross section.}

\item{rv_rate}{Rate parameter for revegetation, passed to \code{revegetate()}.}

\item{nsim}{Number of event hydrographs to run; positive integer.}

\item{niter}{Number of iterations for \code{gbem::gbem()} when running
each hydrograph.}
}
\value{
A stochasim object, containing all the event hydrographs, and
cross sections.
}
\description{
Stochasim Algorithm
}
\examples{
library(distionary)
regime <- hydist_snow(distionary::dst_gev(1000, 3, 0.1), baseflow = 50)
cs <- gbem::cross_section(3, grad = 0.01, d50 = 0.1, d84 = 0.5, roughness = 0.01)
ss <- stochasim(regime, cross_section = cs)
ss_flows(ss)
ss_widths(ss)
}
