% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasim.R, R/stochasim2.R
\name{stochasim}
\alias{stochasim}
\alias{stochasim2}
\title{Stochasim Algorithm}
\usage{
stochasim(x, cross_section, rv_rate = 0.1, niter = 300)

stochasim2(x, y, cross_section, rv_rate = c(0, 0.1), niter = 300)
}
\arguments{
\item{x, y}{Hydrographs, as a list or bare. Recycled to the same length.}

\item{cross_section}{A \code{"cross_section"} object representing a stream's
cross section.}

\item{rv_rate}{Rate parameter for revegetation, passed to \code{revegetate()}.}

\item{niter}{Number of iterations for \code{gbem::gbem()} when running
each hydrograph.}

\item{hydist}{Hydrograph Distribution}

\item{nsim}{Number of event hydrographs to run; positive integer.}
}
\value{
A stochasim object, containing all the event hydrographs, and
cross sections.

A stochasim object, containing all the event hydrographs, and
cross sections.
}
\description{
Stochasim Algorithm

Stochasim Algorithm
}
\examples{
library(distionary)
regime <- hydist_snow(distionary::dst_gev(1000, 3, 0.1), baseflow = 50)
cs <- gbem::cross_section(3, grad = 0.01, d50 = 0.1, d84 = 0.5, roughness = 0.01)
ss <- stochasim(regime, cross_section = cs)
ss_flows(ss)
ss_widths(ss)
# Start with a cross section
cs <- cross_section(3, grad = 0.01, d50 = 0.1, d84 = 0.5, roughness = 0.01)

# Make 1000 rain hydrographs and run stochasim
library(distionary)
dst_rain <- hydist_rain(dst_gev(1000, 3, 0.1), baseflow = 50)
rain <- realise(dst_rain, n = 1000)
ss <- stochasim(hydrographs, cross_section = cs)
plot(ss)
plot(ss, "flows")

# Make 1000 snow hydrographs and run stochasim on both hydrographs.
dst_snow <- hydist_snow(dst_norm(1000, 10^2), baseflow = 50)
snow <- realise(dst_snow, n = 1000)
ss2 <- stochasim2(snow, rain, cross_section = cs)
plot(ss2)
}
